backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncio-atexit==1.0.1
attrs==24.3.0
autogen-agentchat==0.4.3
autogen-core==0.4.3
autogen-ext==0.4.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.11
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.17
distro==1.9.0
dnspython==2.7.0
fastavro==1.10.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==0.3.15
langchain-cohere==0.3.5
langchain-community==0.3.15
langchain-core==0.3.31
langchain-experimental==0.3.4
langchain-openai==0.3.2
langchain-pinecone==0.2.2
langchain-tests==0.3.8
langchain-text-splitters==0.3.5
langgraph==0.2.67
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.0
MarkupSafe==3.0.2
marshmallow==3.25.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.60.1
opentelemetry-api==1.27.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.2.1
protobuf==4.25.6
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.10.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
syrupy==4.8.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
